1.1.0
* init commit 1.1.0

* steps for standalone get repeats function

* An implementation of GetRepeatsForSequence has been added

* Includes some lib prep testing, also fixes the exclusion of last nucleotide bug

* Prepping update for release.

* Prepping update for release.

* Prepping update for release.

* vhangin ubuntu version in github actions

* changing ubuntu version in github actions and did a tool/run-format.sh

* Changing how the library is compiled and adding an example of using the library.

* Adding back in the examples check to the github actions

* Updating run-all examples code

* Updating run-all examples code

+++ src/umatrix.hpp
@@ -45,9 +45,9 @@ typedef struct {
 class UMatrix {
 public:
   // Class variables
-  unsigned long long maxPeriod;      // Max period is included
-  unsigned long long maxInsertions;  // Max insertion is included
-  unsigned long long maxDeletions;   // Max deletion is included
+  unsigned long long maxPeriod;     // Max period is included
+  unsigned long long maxInsertions; // Max insertion is included
+  unsigned long long maxDeletions;  // Max deletion is included
 
   unsigned long long length;         // Number of columns in matrix
   unsigned long long cellsPerColumn; // Number of rows in matrix
@@ -92,7 +92,7 @@ public:
   int *ForwardTraceback(int *traceArray, int windowLength, int row);
 
   void RestartMatrix();
-  void CalculateTraceback(int startColumn);
+  void CalculateTraceback(unsigned long long startColumn);
 
   // Class management
   void CreateMatrix();
